requirements.txt
blinker==1.4
celery==5.0.5
eventlet==0.30.1
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Script==2.0.6
Flask-SocketIO==5.0.1
sentry-sdk[Flask]==0.16.2
flask-redis==0.4.0
gunicorn==20.0.4
requests==2.25.1
# For debug/testing
# ptvsd==4.3.2
# fakeredis==1.4.5
# pytest==6.2.2
# pytest-mock==3.5.1